Initial implementation of kak-debug.
This initial implementation of kak-debug is a Python reimplementation
of kak-dap, but it will not remain that way. The reason is that,
sadly, the design of kak-dap, while interesting as an experiment in
achieving a debug UI with Kakoune buffers, fundamentally flawed. Kakoune
simply doesn't offer a flexible enough feature set (by design) to
make a fully Kakoune debug UI feasible. So, we will shift the approach
to running a TUI application alongside Kakoune that will provide the
stack trace view, variables view, etc, which will give us much more
flexibility in what we can do from a UI standpoint.